Attempt to move role gen to the saas pipeline
